add construct error func
