+++ telematic_system/telematic_apps/web_app/client/src/api/api-states.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import {env} from "../env"
+import { constructError } from './api-utils';
 
 /**
  *@brief Find all states in the US
@@ -12,7 +13,7 @@ const findAllStates = async () => {
     return data;
   } catch (err) {
     
-    return { errCode:  err.response!==undefined && err.response.status !== undefined && err.response.status, errMsg:  err.response !==undefined && err.response.data !== undefined && err.response.data.message !== undefined ? err.response.data.message : (err.response !==undefined ? err.response.statusText :"")}
+    return constructError(err)
   }
 }
