+++ telematic_system/telematic_apps/web_app/client/src/api/api-testing-types.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import {env} from "../env"
+import { constructError } from './api-utils';
 
 /**
  *@brief Find all testing types
@@ -12,7 +13,7 @@ const findAllTestingTypes= async (criteria) => {
         return data;
     } catch (err) {
         
-          return { errCode: err.response!== undefined ? err.response.status: "", errMsg:  err.response !== undefined  && err.response.data !== undefined && err.response.data.message !== undefined ? err.response.data.message : (err.response !== undefined ? err.response.statusText : "")}
+          return constructError(err)
   
     }
 }
